pass data via stdin to support 1.5.9+
JSCS v1.5.9 introduced support for reading via stdin. Unfortunately, its
detection logic is primitive and if stdin is not a TTY (as in
SublimeLinter's case), it will *always* attempt to read from stdin, even
if a filename was passed.

As a workaround, just pass the file via stdin and require JSCS 1.5.9+.
This strategy should be compatible with all future versions.